Finishing user and auth
Still not testing yet

+++ Cloud Computing/controllers/user.js
@@ -4,12 +4,21 @@ const nanoid = require('nanoid');
 const getUserById = (req, res) => {
     const userId = req.params.id;
 
+    if(!userId) {
+        res.status(400).json({
+            status: 'error',
+            message: 'Invalid request. Please provide user id'
+        });
+        return;
+    }
+
     db.query('SELECT * FROM users WHERE id = ?', [userId], (error, results) => {
         if(error){
-            res.status(500).send({
-                message: 'Internal server error, Try agail later',
-                error: error
+            res.status(500).json({
+                status: 'error',
+                message: 'Internal server error. Cannot find user, please try again later',
             });
+            console.log(error);
             return;
         }
 
@@ -32,16 +41,25 @@ const changeUsername = (req, res) => {
     const userId = req.params.id;
     const { newUsername } = req.body;
 
+    if(!userId || !newUsername) {
+        res.status(400).json({
+            status: 'error',
+            message: 'Invalid request. Please provide user id and new username'
+        });
+        return;
+    }
+
     db.query('UPDATE users SET name = ?, WHERE id = ?', [newUsername, userId], (error, results) => {
         if(error){
-            res.status(500).send({
-                message: 'Internal server error, Try agail later',
-                error: error
+            res.status(500).json({
+                status: 'error',
+                message: 'Internal server error. Cannot find user, please try again later',
             });
+            console.log(error);
             return;
         }
 
-        res.status(200).send({
+        res.status(200).json({
             status: 'success',
             message: 'Username change successfully'
         })
@@ -50,18 +68,26 @@ const changeUsername = (req, res) => {
 
 const checkPoints = (req, res) => {
     const userId = req.params.id;
+    if(!userId) {
+        res.status(400).json({
+            status: 'error',
+            message: 'Invalid request. Please provide user id'
+        });
+        return;
+    }
     db.query('SELECT * FROM users WHERE id = ?', [userId], (error, results) => {
         if(error){
-            res.status(500).send({
-                message: 'Internal server error, Try agail later',
-                error: error
+            res.status(500).json({
+                status: 'error',
+                message: 'Internal server error. Cannot find user, please try again later',
             });
+            console.log(error);
             return;
         }
 
         const user = results[0];
 
-        res.status(200).send({
+        res.status(200).json({
             status: 'success',
             points: user.total_points
         })
@@ -71,19 +97,28 @@ const checkPoints = (req, res) => {
 const exchangePoints = (req, res) => {
     const { merchId, merchPoints } = req.body;
     const userId = req.params.id;
+
+    if(!userId || !merchId || !merchPoints) {
+        res.status(400).json({
+            status: 'error',
+            message: 'Invalid request. Please provide user id, merch id, and merch points'
+        });
+        return;
+    }
     db.query('SELECT * FROM users WHERE id = ?', [userId], (error, results) => {
         if(error){
-            res.status(500).send({
-                message: 'Internal server error, Try again later',
-                error: error
+            res.status(500).json({
+                status: 'error',
+                message: 'Internal server error. Cannot find user, please try again later',
             });
+            console.log(error);
             return;
         }
         
         const user = results[0];
 
         if(merchPoints > user.total_points) {
-            res.status(400).send({
+            res.status(400).json({
                 message: 'Points is not enough to exchange this merch'
             });
             return;
@@ -93,31 +128,37 @@ const exchangePoints = (req, res) => {
 
         db.query('UPDATE users SET total_points = ? WHERE id = ?', [total_points, userId], (error, results2) => {
             if(error){
-                res.status(500).send({
-                    message: 'Internal server error, Try again later'
+                res.status(500).json({
+                    status: 'error',
+                    message: 'Internal server error. Cannot update user, please try again later',
                 });
+                console.log(error);
                 return;
             }
 
             db.query('UPDATE merch SET stok = stok - 1 WHERE id = ?', [merchId], (error, results3) => {
                 if(error){
-                    res.status(500).send({
-                        message: 'Internal server error, Try again later'
+                    res.status(500).json({
+                        status: 'error',
+                        message: 'Internal server error. Cannot update merch, please try again later',
                     });
+                    console.log(error);
                     return;
                 }
 
                 const id = nanoid(20)
 
-                db.query('INSERT INTO users_merch_redeem VALUES (?, ?, ?)', [id, userId, merchId], (error, results4) => {
+                db.query('INSERT INTO users_merch_redeem VALUES (?, ?, ?, NOW())', [id, userId, merchId], (error, results4) => {
                     if(error){
-                        res.status(500).send({
-                            message: 'Internal server error, Try again later'
+                        res.status(500).json({
+                            status: 'error',
+                            message: 'Internal server error. Cannot insert merch redeem by user, please try again later',
                         });
+                        console.log(error);
                         return;
                     }
 
-                    res.status(200).send({
+                    res.status(200).json({
                         status: 'success',
                         message: 'Points is successfully exchange'
                     })
@@ -129,20 +170,29 @@ const exchangePoints = (req, res) => {
 }
 
 const addNotification = (req, res) => {
-    const { userId, title, message, date } = req.body;
+    const { userId, title, message} = req.body;
+    if(!userId || !title || !message){
+        res.status(400).json({
+            status: 'error',
+            message: 'Invalid request. Please provide user id, title, and message'
+        });
+        return;
+    }
 
     const id = nanoid(20);
 
-    db.query('INSERT INTO users_notifications VALUES (?, ?, ?, ?, ?)', 
+    db.query('INSERT INTO users_notifications VALUES (?, ?, ?, ?, NOW())', 
     [id, userId, title, message, date], (error, results) => {
-        if(error) {
-            res.status(500).send({
-                message: 'Internal server error, try again later'
+        if(error){
+            res.status(500).json({
+                status: 'error',
+                message: 'Internal server error. Cannot insert user notification, please try again later',
             });
+            console.log(error);
             return;
         }
         
-        res.status(201).send({
+        res.status(201).json({
             status: 'success',
             message: 'Notification is created successfully'
         })
@@ -151,16 +201,25 @@ const addNotification = (req, res) => {
 
 const getNotificationByUserId = (req, res) => {
     const userId = req.params.id;
+    if(!userId){
+        res.status(400).json({
+            status: 'error',
+            message: 'Invalid request. Please provide user id'
+        });
+        return;
+    }
 
     db.query('SELECT * FROM users_notifications WHERE user_id = ?',  [userId], (error, results) => {
         if(error){
-            res.status(500).send({
-                message: 'Internal server error, try again later'
+            res.status(500).json({
+                status: 'error',
+                message: 'Internal server error. Cannot find user notification, please try again later',
             });
+            console.log(error);
             return;
         }
 
-        res.status(200).send({
+        res.status(200).json({
             status: 'success',
             data: results
         })
@@ -169,15 +228,24 @@ const getNotificationByUserId = (req, res) => {
 
 const deleteNotificationById = (req, res) => {
     const notifId = req.params.id;
+    if(!notifId){
+        res.status(400).json({
+            status: 'error',
+            message: 'Invalid request. Please provide notification id'
+        });
+        return;
+    }
     db.query('DELETE FROM user_notifications WHERE id = ?', [notifId], (error, results) => {
-        if(error) {
-            res.status(500).send({
-                message: 'Internal server error, try again later'
+        if(error){
+            res.status(500).json({
+                status: 'error',
+                message: 'Internal server error. Cannot delete user notification, please try again later',
             });
+            console.log(error);
             return;
         }
 
-        res.status(200).sensd({
+        res.status(200).json({
             status: 'success',
             message: 'notifications is successfully delete'
         })
@@ -186,14 +254,44 @@ const deleteNotificationById = (req, res) => {
 
 const getAllMerch = (req, res) => {
     db.query('SELECT * FROM merch', (error, results) => {
-        if(error) {
-            res.status(500).send({
-                message: 'Internal server error, try again later'
+        if(error){
+            res.status(500).json({
+                status: 'error',
+                message: 'Internal server error. Cannot get data from merch, please try again later',
             });
+            console.log(error);
             return;
         }
 
-        res.status(200).send({
+        res.status(200).json({
+            status: 'success',
+            data: results
+        });
+    });
+}
+
+const getMerchRedeemedByUserId = (req, res) => {
+    const userId = req.params.id;
+    if(!userId){
+        res.status(400).json({
+            status: "error",
+            message: "Invalid request. Please provide user id"
+        });
+        return;
+    }
+
+    db.query('SELECT b.name as name, b.image as image, a.date as date FROM users_merch_redeem a JOIN merch b ON a.merch_id = b.id JOIN users c ON a.user_id = c.id WHERE a.user_id = ?',
+    [userId], (error, results) => {
+        if(error){
+            res.status(500).json({
+                status: 'error',
+                message: 'Internal server error. Cannot get data, please try again later',
+            });
+            console.log(error);
+            return;
+        }
+
+        res.status(200).json({
             status: 'success',
             data: results
         });
@@ -208,5 +306,6 @@ module.exports = {
     addNotification,
     getNotificationByUserId,
     deleteNotificationById,
-    getAllMerch
+    getAllMerch,
+    getMerchRedeemedByUserId
 };
